Fixed some camera jitter
Moved the spider rotation smoother rotation matcher to fixed update and the jitter is gone.  However doing so causes the framerate of the camera to be the same as the physics framerate.  Tradeoff is only setting the mainbody rotation to be in fixed update.  So there is still jitter when moving parallel to an object but none really anywhere else.